Multi-llm/requirements.txt
openai>=1.0.0
anthropic>=0.7.0
google-generativeai>=0.3.0
groq>=0.3.0
requests>=2.31.0
python-dotenv>=1.0.0
pytest>=7.0.0